+++ app/lab/murajaah-at-taisir/components/AyahTextHighlight.tsx
@@ -1,0 +1,43 @@
+type AyahTextHighlightProps = {
+  ayahText: string;
+  isSelected: boolean;
+  isSurahAlfatiha?: boolean;
+};
+
+export function AyahTextHighlight({
+  ayahText,
+  isSelected,
+  isSurahAlfatiha = false,
+}: AyahTextHighlightProps) {
+  const BISMILLAH = "بِسۡمِ ٱللَّهِ ٱلرَّحۡمَـٰنِ ٱلرَّحِیمِ ";
+  let displayText = ayahText;
+
+  if (!isSurahAlfatiha) {
+    displayText = ayahText.replace(BISMILLAH, "");
+  }
+
+  function getFirstArabicWord(text: string) {
+    const words = text.split(" ");
+    const isTooLong = words[0].split("").length > 6 || words.length < 6;
+    return words.slice(0, isTooLong ? 1 : 2).join(" ");
+  }
+
+  function removeFirstArabicWord(text: string) {
+    const words = text.split(" ");
+    const isTooLong = words[0].split("").length > 6 || words.length < 6;
+    return words.slice(isTooLong ? 1 : 2).join(" ");
+  }
+
+  if (!isSelected) {
+    return <span className="opacity-[0.025]">{displayText.trim()}</span>;
+  }
+
+  return (
+    <span className="bg-[#f1efdf] hover:[&>span]:opacity-100 [&>span]:transition-opacity">
+      <span>{getFirstArabicWord(displayText).trim()}</span>{" "}
+      <span className="opacity-[0.025]">
+        {removeFirstArabicWord(displayText).trim()}
+      </span>
+    </span>
+  );
+}