+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Improved ayah display logic in murajaah-at-taisir (see details in [app/lab/murajaah-at-taisir/CHANGELOG.md](app/lab/murajaah-at-taisir/CHANGELOG.md)).
+- Improved ayah display logic and layout shifting in murajaah-at-taisir, with refactored code for better readability. See details in [app/lab/murajaah-at-taisir/CHANGELOG.md](app/lab/murajaah-at-taisir/CHANGELOG.md)).
 
 ## [v0.10.3] - 2025-11-5